Use Ramsey's UUID library when generating task IDs
The UUID function in Fuel core does not use a good source of entropy,
instead using a faster Mersenne Twister algorithm (`mt_rand`) which
seems to collide a lot.

We've got 48,059,104 unique IDs in production and we've seen over 1,600
collisions in the last 6 months. Ramsey's UUID library meanwhile uses
the `random_bytes` function in PHP7+ (or a fallback using the
paragonie/random_compat library), both of which use OpenSSL randomness
which is more random.

This should reduce the number of effective collisions to zero, however
the ultimate fix would involve regenerating the ID when it collides in
the database.